requirements.txt
openai-whisper
pyAudioAnalysis
librosa
soundfile
numpy
scipy
torch
torchaudio
